+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Dependency Updates
 
 * lombok from `1.18.42` to `1.18.44`
+* Removed morse-code-translator dependency
 
 ## [1.0.0-beta.1] - 2026-03-08